refactor: reorganizing dir structure
