+++ src/components/CodeBrackets/index.scss
@@ -28,7 +28,9 @@ $total: 10;
   $scale: random(2) - 0.4;
 
   .svg:nth-child(#{$i}) {
+    stroke-width: '19px';
     stroke: #{$nth};
+    fill: #{$nth};
     opacity: 0.4;
     left: random(120) * 1% - 20;
     animation: raise#{$i} 6 + random(150) + s linear infinite;
@@ -40,7 +42,7 @@ $total: 10;
     @keyframes raise#{$i} {
       to {
         bottom: 150vh;
-        transform: scale(0.3 * $i - 0.7) rotate(random(10) + deg);
+        transform: scale(0.1 * $i - 0.7) rotate(random(10) + deg);
       }
     }
   }
