--- src/components/Home/index.scss
@@ -1,73 +1,0 @@
-@import '../../variables.scss';
-
-.home-page {
-  h1 {
-    font-family: 'Coolvetica';
-    font-weight: 400;
-    &::before {
-      content: '<h1>';
-      font-family: 'La Belle Aurore';
-      color: $primary-color-light;
-      font-size: 18px;
-      position: absolute;
-      margin-top: -40px;
-      left: 15px;
-      opacity: 0.6;
-    }
-    &::after {
-      content: '<h1/>';
-      font-family: 'La Belle Aurore';
-      color: $primary-color-light;
-      font-size: 18px;
-      position: absolute;
-      margin-top: 18px;
-      margin-left: 20px;
-      animation: fadeIn 1s 1.7s backwards;
-      opacity: 0.6;
-    }
-    img {
-      height: 65px;
-      opacity: 0;
-      animation: rotateIn 1s linear both;
-      animation-delay: 1.4s;
-      &.letter-t {
-        margin-right: -35px;
-        display: inline-block;
-      }
-      &.letter-b {
-        margin-left: 10px;
-        display: inline-block;
-      }
-    }
-  }
-}
-
-h2 {
-  color: #8d8d8d;
-  margin-top: 20px;
-  font-weight: 400;
-  font-size: 11px;
-  font-family: sans-serif;
-  letter-spacing: 3px;
-  animation: fadeIn 1s 1.8s backwards;
-}
-
-.flat-button {
-  color: $primary-color-light;
-  font-size: 13px;
-  font-weight: 400;
-  letter-spacing: 4px;
-  font-family: sans-serif;
-  text-decoration: none;
-  padding: 10px 18px;
-  border: 1px solid $primary-color-light;
-  margin-top: 25px;
-  animation: fadeIn 1s 1.8s backwards;
-  white-space: nowrap;
-  align-self: flex-start;
-
-  &:hover {
-    background: $alt-color-light;
-    color: $text-color-dark;
-  }
-}
