+++ src/components/HomePicture/index.scss
