+++ src/components/pages/About/index.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import '../../../variables.scss';
 .stage-cube-cont {
   width: 50%;
   height: 100%;
