+++ src/components/pages/Contact/index.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import '../../../variables.scss';
 
 .contact-form {
   width: 100%;
